Fixed point tiles and min size limit

+++ tiles/tiles.hpp
@@ -15,7 +15,7 @@ public:
 	typedef unsigned int Tile;
 	typedef unsigned int Pos;
 
-	typedef float Cost;
+	typedef long int Cost;
 	typedef int Oper;
 	enum {
 		Nop = -1,
